+++ SandboxGame/source/CCustomEngineListener.cpp
@@ -505,9 +505,6 @@ E_RESULT_CODE CCustomEngineListener::OnStart()
 
 	mpResourceManager->Load<IPrefabsManifest>("ProjectResources/PrefabsCollection.manifest");
 
-	auto pPrefabsRegistry = CreatePrefabsRegistry(mpEngineCoreInstance->GetSubsystem<IResourceManager>(), mpEngineCoreInstance->GetSubsystem<IFileSystem>(), mpWorld, result);
-	auto pEntity = pPrefabsRegistry->Spawn("TestPrefab");
-
 	return RC_OK;
 }